code refactoring, removed old codes and some UI fixes
