+++ modules/models/topicsList.go
@@ -1,0 +1,17 @@
+package models
+
+// TopicsList return a list of topics/categories
+func TopicsList() Topics {
+	topics := Topics{
+		"society":       TopicIdentity{"y", "社会"},
+		"international": TopicIdentity{"w", "国際"},
+		"business":      TopicIdentity{"b", "ビジネス"},
+		"politics":      TopicIdentity{"p", "政治"},
+		"entertainment": TopicIdentity{"e", "エンタメ"},
+		"sports":        TopicIdentity{"s", "スポーツ"},
+		"technology":    TopicIdentity{"t", "テクノロジー"},
+		"pickup":        TopicIdentity{"ir", "ピックアップ"},
+	}
+
+	return topics
+}
